@@ -1,4 +1,4 @@
-// swift-tools-version: 5.8
+// swift-tools-version: 5.9
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -11,7 +11,7 @@ let package = Package(
         // .package(url: /* package url */, from: "1.0.0"),
         .package(
             url: "https://github.com/swift-cloud/Compute",
-            from: "2.17.0"
+            from: "2.18.0"
         ),
 //        .package(
 //            url: "https://github.com/johnsundell/ink.git",
@@ -21,9 +21,9 @@ let package = Package(
             url: "https://github.com/yumemi-inc/fake-fortune-telling",
             from: "0.2.5"
         ),
-        .package(url: "https://github.com/apple/swift-openapi-generator", from: "0.2.1"),
-        .package(url: "https://github.com/apple/swift-openapi-runtime", from: "0.2.1"),
-        .package(url: "https://github.com/novr/swift-openapi-compute", from: "0.0.1"),
+        .package(url: "https://github.com/apple/swift-openapi-generator", from: "1.2.0"),
+        .package(url: "https://github.com/apple/swift-openapi-runtime", from: "1.3.2"),
+        .package(url: "https://github.com/novr/swift-openapi-compute", from: "0.1.2"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
