@@ -1,10 +1,15 @@
 // Generated by swift-openapi-generator, do not modify.
 @_spi(Generated) import OpenAPIRuntime
 #if os(Linux)
-@preconcurrency import Foundation
+@preconcurrency import struct Foundation.URL
+@preconcurrency import struct Foundation.Data
+@preconcurrency import struct Foundation.Date
 #else
-import Foundation
+import struct Foundation.URL
+import struct Foundation.Data
+import struct Foundation.Date
 #endif
+import HTTPTypes
 extension APIProtocol {
     /// Registers each operation handler with the provided transport.
     /// - Parameters:
@@ -13,9 +18,9 @@ extension APIProtocol {
     ///   request handlers.
     ///   - configuration: A set of configuration values for the server.
     ///   - middlewares: A list of middlewares to call before the handler.
-    public func registerHandlers(
+    internal func registerHandlers(
         on transport: any ServerTransport,
-        serverURL: URL = .defaultOpenAPIServerURL,
+        serverURL: Foundation.URL = .defaultOpenAPIServerURL,
         configuration: Configuration = .init(),
         middlewares: [any ServerMiddleware] = []
     ) throws {
@@ -26,13 +31,19 @@ extension APIProtocol {
             middlewares: middlewares
         )
         try transport.register(
-            { try await server.post_hyphen_my_hyphen_fortune(request: $0, metadata: $1) },
+            {
+                try await server.post_hyphen_my_hyphen_fortune(
+                    request: $0,
+                    body: $1,
+                    metadata: $2
+                )
+            },
             method: .post,
-            path: server.apiPathComponentsWithServerPrefix(["my_fortune"]),
-            queryItemNames: []
+            path: server.apiPathComponentsWithServerPrefix("/my_fortune")
         )
     }
 }
+
 fileprivate extension UniversalServer where APIHandler: APIProtocol {
     /// Get fortune for given person
     ///
@@ -40,13 +51,20 @@ fileprivate extension UniversalServer where APIHandler: APIProtocol {
     ///
     /// - Remark: HTTP `POST /my_fortune`.
     /// - Remark: Generated from `#/paths//my_fortune/post(post-my-fortune)`.
-    func post_hyphen_my_hyphen_fortune(request: Request, metadata: ServerRequestMetadata) async throws -> Response {
+    func post_hyphen_my_hyphen_fortune(
+        request: HTTPTypes.HTTPRequest,
+        body: OpenAPIRuntime.HTTPBody?,
+        metadata: OpenAPIRuntime.ServerRequestMetadata
+    ) async throws -> (HTTPTypes.HTTPResponse, OpenAPIRuntime.HTTPBody?) {
         try await handle(
             request: request,
-            with: metadata,
+            requestBody: body,
+            metadata: metadata,
             forOperation: Operations.post_hyphen_my_hyphen_fortune.id,
-            using: { APIHandler.post_hyphen_my_hyphen_fortune($0) },
-            deserializer: { request, metadata in
+            using: {
+                APIHandler.post_hyphen_my_hyphen_fortune($0)
+            },
+            deserializer: { request, requestBody, metadata in
                 let headers: Operations.post_hyphen_my_hyphen_fortune.Input.Headers = .init(
                     API_hyphen_Version: try converter.getOptionalHeaderFieldAsURI(
                         in: request.headerFields,
@@ -57,36 +75,51 @@ fileprivate extension UniversalServer where APIHandler: APIProtocol {
                 )
                 let contentType = converter.extractContentTypeIfPresent(in: request.headerFields)
                 let body: Operations.post_hyphen_my_hyphen_fortune.Input.Body
-                if try contentType == nil
-                    || converter.isMatchingContentType(received: contentType, expectedRaw: "application/json")
-                {
-                    body = try converter.getRequiredRequestBodyAsJSON(
+                let chosenContentType = try converter.bestContentType(
+                    received: contentType,
+                    options: [
+                        "application/json"
+                    ]
+                )
+                switch chosenContentType {
+                case "application/json":
+                    body = try await converter.getRequiredRequestBodyAsJSON(
                         Components.Schemas.MyFortuneRequest.self,
-                        from: request.body,
-                        transforming: { value in .json(value) }
+                        from: requestBody,
+                        transforming: { value in
+                            .json(value)
+                        }
                     )
-                } else {
-                    throw converter.makeUnexpectedContentTypeError(contentType: contentType)
+                default:
+                    preconditionFailure("bestContentType chose an invalid content type.")
                 }
-                return Operations.post_hyphen_my_hyphen_fortune.Input(headers: headers, body: body)
+                return Operations.post_hyphen_my_hyphen_fortune.Input(
+                    headers: headers,
+                    body: body
+                )
             },
             serializer: { output, request in
                 switch output {
                 case let .ok(value):
                     suppressUnusedWarning(value)
-                    var response = Response(statusCode: 200)
+                    var response = HTTPTypes.HTTPResponse(soar_statusCode: 200)
                     suppressMutabilityWarning(&response)
+                    let body: OpenAPIRuntime.HTTPBody
                     switch value.body {
                     case let .json(value):
-                        try converter.validateAcceptIfPresent("application/json", in: request.headerFields)
-                        response.body = try converter.setResponseBodyAsJSON(
+                        try converter.validateAcceptIfPresent(
+                            "application/json",
+                            in: request.headerFields
+                        )
+                        body = try converter.setResponseBodyAsJSON(
                             value,
                             headerFields: &response.headerFields,
                             contentType: "application/json; charset=utf-8"
                         )
                     }
-                    return response
-                case let .undocumented(statusCode, _): return .init(statusCode: statusCode)
+                    return (response, body)
+                case let .undocumented(statusCode, _):
+                    return (.init(soar_statusCode: statusCode), nil)
                 }
             }
         )
